refactor: move resource-assertions to test/util
