+++ packages/opentelemetry-node/test/NodeTracerProvider.test.ts
@@ -24,7 +24,7 @@ import {
 } from '@opentelemetry/core';
 import { AsyncHooksScopeManager } from '@opentelemetry/scope-async-hooks';
 import { Span } from '@opentelemetry/tracing';
-import { assertTelemetrySDKResource } from '@opentelemetry/resources';
+import { Resource, TELEMETRY_SDK_RESOURCE } from '@opentelemetry/resources';
 import * as assert from 'assert';
 import * as path from 'path';
 import { ScopeManager } from '../../opentelemetry-scope-base/build/src';
@@ -160,16 +160,17 @@ describe('NodeTracerProvider', () => {
       assert.deepStrictEqual(span.attributes, defaultAttributes);
     });
 
-    it('should assign a telemetry sdk resource to span', () => {
+    it('should assign resource to span', () => {
       provider = new NodeTracerProvider({
         logger: new NoopLogger(),
       });
       const span = provider.getTracer('default').startSpan('my-span') as Span;
       assert.ok(span);
-      assertTelemetrySDKResource(span.resource, {
-        language: 'nodejs',
-        name: 'opentelemetry',
-      });
+      assert.ok(span.resource instanceof Resource);
+      assert.equal(
+        span.resource.labels[TELEMETRY_SDK_RESOURCE.LANGUAGE],
+        'nodejs'
+      );
     });
   });
