+++ packages/opentelemetry-resources/src/index.ts
@@ -16,4 +16,3 @@
 
 export { Resource } from './Resource';
 export * from './constants';
-export * from './resource-assertions';
